+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 Changelog
 =========
+## Unreleased
+### Added
+- FlashMessages filter
+
 ## 1.0.0-rc.2 - 2017-12-06
 ### Added
 - Card asset bundle